Fix array-of-struct schema evolution failing to cast when containsNull is false.
There is no user-facing API involved in this PR fix.

Problem:

When the column is constructed by spark sql function `array(struct(…))`, all the struct fields are not nullable, and `containsNull` in that arrayType is false. Because castIfNeeded is called recursively to cast every array structType elements, which generates nullable expressions, ArrayTransform will generate an ArrayType with containsNull as true. Thus, an AnalysisException is thrown saying that it cannot resolve the two types because of the inconsistency in `containsNull` between the arrayType generated by ArrayTransform and the one being casted to when the schema evolution should be resolved properly.

Fix:

We need to make `containsNull` in `to` as true so casting won't fail in the above case.

Unit test.

GitOrigin-RevId: 38163dfc9f6b47ec70851507a70bec2e8e82f0fb