Calculate MERGE numSourceRows from findTouchedFiles
Calculate the MERGE metric `numSourceRows` from the first job (findTouchedFiles) instead of from subsequent jobs.

The problem is that numSourceRows as captured in Job 2 doesn't always get captured because Job 2 doesn't always run. So instead, we move the calculation of numSourceRows to Job 1 (preserving the existing metric, but calculating it in a different way), but we then re-add the calculation in Job 2 (if it runs) under a different name. In the diff it looks like renaming the metric, but it's really a move + re-add.

Changes:
* Renamed the existing metric from the other jobs to `numSourceRowsInSecondScan`.
* Added calculation of `numSourceRows` to `findTouchedFiles`.
* Added a check for mismatches between those two values, enabled only by a config that is disabled by default. This will help us to flag nondeterministic sources.

Unit tests. New test for the config that detects nondeterministic sources.

GitOrigin-RevId: 2d92c969a4077449b85e661c9044fac9aa095759