Fail MERGE queries that are adding void columns
Void (`NullType`) columns are not properly supported in MERGE command. Specifically, the first merge can successfully add a void column, but subsequent merges will fail with a mismatch error. Until proper void column support, we prefer to fail any query that tries to add a void column and return an error message that instructs users to explicitly add a type.

Note that different kind of writes can still add void columns to tables, and customers will not be able to merge into these tables until they provide a type for all columns.

Added new unit tests that check that void column or nested fields in merge are not allowed.

GitOrigin-RevId: c7d4129ecb8c27b399658eaba76f8a1976010c3a