Tag Delta Metadata Queries for Logging using a generalised frame-like tagging mechanism
This PR adds a general utility object that allows "tagging" queries that are executed in an enclosing block with support for nested blocks. These "tags" are read by the usage logger/query profile logger and all parent/ancestor tags of the queries are added to the `QueryProfile`. This mechanism is implemented as a utility object `QueryTagger` (which extends the `ThreadLocalTagger` trait that contains the core logic).

The tag is currently used for tagging Delta Metadata Queries but the tags can be easily extended to support tagging of say, queries generated by MERGE INTO.

New unit tests + existing tests.

GitOrigin-RevId: 403432a9baa0cda4d414acb8f57690cdd5b8dd19