Don't update snapshot if within the same request
Allow SnapshotManagement.update to take in a timestamp, which if specified can be used to identify if the snapshot has already been updated in the duration of the current request. This enables us to save a listFrom call to the underlying logstore, since previously DeltaTableV2 could potentially trigger up to 2 LISTs - one when fetching the table (if it wasn't in the cache), and 1 when evaluating the `lazy val snapshot`. With the change, we no-op during the second update if the table was fetched from the logstore.

GitOrigin-RevId: 01d1bccecb2fd1d4446102ecad561c1757c1049e