Avoid unnecessary struct() creation during state reconstruction.
This PR fixes a performance regression in Delta state reconstruction due to bad query planning: When the Delta checkpoint schema contains `add.stats_parsed` column instead of `add.stats` (because JSON stats are disabled while struct stats are enabled), we must synthesize the json stats which in turn requires a rewrite of the `add` struct. Further, the struct was already rewritten before that, in order to canonicalize `add.path`. All of these rewrites were replicated at every access of any field of the final `add` struct, causing the `to_json(add.stats_parsed)` projection to run ten times instead of just once. Path canonicalization also ran multiple times, but the JSON overhead dwarfed it.

This behavior is a definite shortcoming in the query optimizer's handling of nested fields, but there is no immediately obvious fix for it.

The workaround is to rework the state reconstruction query, so that the `add` struct rewrite that applies stats construction and path canonicalization happens at the last possible moment -- just before converting to `Dataset[SingleAction]` -- so that the optimizer has no opportunity to replicate the projections. The result is actually faster than the pre-regression code, because even the original state reconstruction still evaluated the struct twice.

Many existing unit tests cover correctness.

GitOrigin-RevId: f4fa51a0a900f8a8b892ed4129fad69932181296