Multipart checkpoints
Resolves delta-io/delta#837

Adds a config that sets the maximum number of actions to include in a single file of a multipart checkpoint. If the total actions of a snapshot is larger than this, the checkpoint will be split up into multiple parts.

It is disabled by default so the behavior is purely opt-in, as most users probably don't know about actions or how to set this value. In the future it could be set to some sane default value if there's some consensus on what it should be.

Closes delta-io/delta#946

Signed-off-by: Venki Korukanti <venki.korukanti@gmail.com>
GitOrigin-RevId: b57610d842256bdacd10a6986a71758de777f48d