Do not overwrite numCopiedRows for DELETE commands
During the collection of Metrics for an OptimisticTransaction, the value for "numCopedRows" was overwritten. This was done to fix errors in metric collection code.

There are already sufficiently many tests for the different metrics

The operational metrics for DELETE were wrong and differed from the correct values in the usage logs. This has been fixed.

GitOrigin-RevId: 73fab1d54e094c897857a28b7fa5583f102411d7