Make createdTime in metadata as None by default
Today if we use JsonUtils to deserialize an Metadata object that doesn't have createdTime field, Jackson will fill the current system timestamp. This is not desired and is not consistent with the behavior when using Spark to read. This inconsistency may cause issues in future. Hence we change the default value of createdTime to None and when a caller create a Metadata object, it should set it explicitly.

In Delta protocol the createdTime is optional, and thus its totally valid for it to be empty.

Unit test. the same unit test failed without this fix.

GitOrigin-RevId: 04989db590694d3057b948d94e74dcdd81f80ad6