Delta Lake CDF - Stream API
This PR adds CDF + Streaming functionality, as part of the ongoing CDF project.
The bulk of this PR is
a) adding `DeltaSourceCDCSupport`. This PR adds on the necessary CDF functionality to DeltaSource.
b) updating DeltaSource to use the various `DeltaSourceCDCSupport` APIs when CDF is enabled
c) adding a new test suite

Closes delta-io/delta#1154

GitOrigin-RevId: e7e8f6d48f99a63e7c5d35a5d0173a9cc26cf274