Minor refactor to DeltaErrors and DeltaErrorsSuite
GitOrigin-RevId: ceacc3a6239f36e76cc8cf4f3447285bb06fc6a0

CDF + Vacuum tests

This PR adds two tests testing CDF + VACUUM integration.

Closes delta-io/delta#1177

GitOrigin-RevId: f2f7b187cb3cc78c267d378eaf3d0657a56241d9

Adds miscellaneous (e.g. end-to-end workload, and CDCReader) tests for CDF.

Resolves delta-io/delta#1178

GitOrigin-RevId: 5c7da4ff9413d84e73137a80673872065de8267b

Minor refactor to UpdateCommand

GitOrigin-RevId: ab3fcbe0522aa194ac0781730a50112655d5c7ec

 Fixes #348 Support Dynamic Partition Overwrite

The goal of this PR to to support dynamic partition overwrite mode on writes to delta.

To enable this on a per write add `.option("partitionOverwriteMode", "dynamic")`. It can also be set per sparkSession in the SQL Config using `.config("spark.sql.sources.partitionOverwriteMode", "dynamic")`.

Some limitations of this pullreq:
Dynamic partition overwrite mode in combination with replaceWhere is not supported. If both are set this will result in an error.
The SQL `INSERT OVERWRITE` syntax does not yet support dynamic partition overwrite. For this more changes will be needed to be made to `org.apache.spark.sql.delta.catalog.DeltaTableV2` and related classes.

Fixes delta-io/delta#348
Closes delta-io/delta#371

Signed-off-by: Allison Portis <allison.portis@databricks.com>
GitOrigin-RevId: 5b01e5b04e573dabe91ac2d71991a127617b8038

Add Checkpoint + CDF test.

This PR adds a test to ensure CDC fields are not included in the checkpoint file.

Resolves delta-io/delta#1180

GitOrigin-RevId: d4a7b8bc4d1a79d30806ff18c5b507f6edcd964c