Cast Expression first then perform array functions in Merge Schema Evolution
We introduced schema evolution for ArrayType to cast each array element into the specified type. However, the fromExpression can be an array function, for example, array_union. In this case, when we apply the transformation on array elements, the array function will also be evaluated for every elements. To avoid this cost, we need to perform the transformation for underlying expression first and then perform the array function.

Unit tests.

GitOrigin-RevId: 11b53b6b421445f4c0fac5447e86550e28611ba7