[ZOrderBy OSS] Add zOrderBy column names to DeltaOperation
(Cherry-pick delta-io/delta#1197)

Add zOrderBy columns to DeltaOperation `Optimize`. This will help log the zOrderBy columns in Delta table history

Added test for checking the Delta history once the zOrderBy command is complete
Added test to verify the zOrderBy command output. As part of this test a bug was found where we are assigning the number of zCubes constructed as 1. It should be equal to the number of partitions.

Close delta-io/delta#1197

GitOrigin-RevId: b4152a73a40d105d604685633dfe0ae8b6a50456