Enhance CDF start timestamp supprot to handle out-of-range timestamp in stream workload timestamp in stream workload
This is a change to follow on the work 786abc7e34f8732a915852310360f3c811cb77ec, and it enables to handle start timestamp out of range in stream workload. Without this change, streaming workload with start timestamp beyond latest commit version throws an error.

The changed behavior is protected under the config DeltaSQLConf.DELTA_CDF_ALLOW_OUT_OF_RANGE_TIMESTAMP which in default is disabled, so it only takes effect when user opts in.

The key change in this patch modifies `getStartingVersion` in DeltaSource.scala so that it won't throw exception if the config is enabled.

Added a new unit test to verify the change

GitOrigin-RevId: 75a9c3fb1a31d9324f6f0c428d77c5e343ccfe90