Tests for operation metrics for CDF + Merge command
This PR adds a new test suit skeleton for tests, and adds 1 concrete test.

This PR adds a new test suite trait `MergeIntoMetricsBase` which is extended only by `trait DescribeDeltaHistorySuiteBase`.

`MergeIntoMetricsBase` is only the skeleton of a test suite - it adds all of the necessary helper methods, but doesn't actually implement all of the tests. Instead, it adds only 1 test. We will add more tests in a followup PR.

GitOrigin-RevId: a0557c9c239d1de8e209e2e6461bb9959f85860e