Block writes that contain CDF + Drop/Rename + FileActions
This PR block commits if
1. table has CDC enabled and there are `FileActions` to write
2. table has column mapping enabled and there is a column mapping related metadata action (e.g. DROP, RENAME, upgrade)

We do this because the current semantics is undefined. So, we block it for now, and can un-block it in the future once we define the expected output.

Please note: under current public APIs, this scenario is not possible. e.g. during a DROP or RENAME operation, only metadata is changes, so no `FileAction`s are committed. Nonetheless, we want to future proof this.

This block occurs during `OptimisticTransactionImpl::prepareCommit`

At a high level, this PR solves the given problem by solving 3 smaller problems

- what is the mapping of physical -> logical names for a given schema? For this we add `DeltaColumnMapping::getPhysicalNameFieldMap`
- given a new schema and an existing schema, detect a drop/rename column operation using the map from 1). For this we add `DeltaColumnMapping::isDropColumnOperation` and `DeltaColumnMapping::isRenameColumnOperation`
- Using the helper method from 2, implement the desired blocking algorithm. For this we add `OptimisticTransaction::performCdcColumnMappingCheck`.

For each of the 3 sub-problems described above, we add a unit test to DeltaColumnMappingSuite.

Closes delta-io/delta#1211

GitOrigin-RevId: 5ea6bc8c6271d2d82b79c8cf462a510d9607546a