Avoid Persisting NumRecord of RemoveFile to Checkpoint
(Cherry-pick of delta-io/delta#1230)

This PR avoids persisting NumRecord of RemoveFile action to checkpoint by removing this attribute from the constructor of RemoveFile object.

Resolves delta-io/delta#1229.

This PR ensures that `RemoveFile.numRecords` field is not written out to the delta checkpoint.

We write out a checkpoint, and read it back as parquet, and ensure that its schema does not contain `numRecords`.

No.

Closes delta-io/delta#1230.

GitOrigin-RevId: 518e46c0622cca4277729e9e6e7ebb08452619f3