Improve error messages for char/varchar type length limitation check in delta check constraints
Improved the error message for varchar type length checks by including the maximum allowed length as well as the expression being evaluated. This should give the end user a better idea of where the error occurred.

Modified an existing unit test and verified both the expected max length as well as the expression being evaluated both show up in the error message.

GitOrigin-RevId: 7c36df8bdc89d7a62b6c5057723dc46322e6cfb9