Fix broken link in PROTOCOL.md file
(Cherry-pick of delta-io/delta#1251)

There was a broken link in `PROTOCOL.md` file. It is the link titled `Delta Sink for Apache Spark's Structured Streaming` inside **Transaction Identifiers** section.

The fix is a simple change in the URL to point to the right source file.

It is a markdown update hence no tests included.

No.

Closes delta-io/delta#1251

Signed-off-by: Venki Korukanti <venki.korukanti@gmail.com>
GitOrigin-RevId: 65122335b90cb7820035a669d19cd9438b314988