[Delta] Accept LogStore conf keys with and without the "spark." prefix
This PR adds support for both `spark.` prefixed and non-`spark.` prefixed keys for LogStore confs (class and scheme.) This is to maintain compatibility across the delta ecosystem.

Originally, I did this similarly to https://github.com/delta-io/connectors/pull/370 where we "normalize" the `SparkConf`, but I think this is less safe than the way I've chosen to do it, because we would need to update all spark configurations that might be accessed from anywhere else in the code. For example, `DelegatingLogStore` uses `SparkEnv.get.conf`.

Instead, we alter our access when looking up keys to check for both spark-prefixed and non-spark-prefixed keys.

Changes in this PR:
- adds `LogStoreProvider. getLogStoreConfValue` which checks for both keys
- adds `verifyLogStoreConfs` which checks for discrepancies between the two keys
- refactors part of `DelegatingLogStoreSuite` into `LogStoreProviderSuite`

Adds tests to `LogStoreProviderSuite` and `DelegatingLogStoreSuite`

GitOrigin-RevId: 2ea19970e3fc40e8f2b06cb70f9c27197f399b08