[Delta] Metric tests for merge
(Cherry-pick of d2804cb)

This PR introduces tests for metrics in Merge. These added tests revealed multiple bugs in our MergeIntoCommand metric calculation. These metrics are further outlined below. This PR also fixes 3 of these issues. The remaining failing tests are ignored in this PR so that we can merge them. When the metrics are fixed in the future we will reenable these tests.

This PR adds tests and some fixes. The fixes are tested by the new tests.

- `numOutputRows`: failing when CDF enabled. Fixed in this PR.
- `numSourceRows`: only with an empty target. Fix in this PR. (non-regression)
- `numTargetRowsDeleted`: failing " delete-only without join", "delete-only with duplicate". Fixed in this PR. (non-regression)
    - The first one was due to the cross-product join bug for Merge CDC. However the fix for duplicate matches also corrects this result.

- `numTargetFilesAdded`: only when `partitioned=false` (non-regression) (expected: 0, found: 1)
- `numTargetRowsCopied`: our implementation seems to return the number of rows persisted (without regards to whether they were physically rewritten or not) with some additional nuance. Not sure of an easy fix for this at the moment. (non-regression)

GitOrigin-RevId: ee67a87fdff3f13c9401d5814b5308c9260d404f