Add and use FileStatus overloads for FileNames utility methods
Today's Delta code is littered with calls such as `isDeltaFile(fileStatus.getPath)` -- which is a bit annoying because most such calls are in the context of file listing, which always returns `FileStatus` rather than `Path`. Especially annoying in the context of collection operations, such as:
```scala
listing
  .filter(f => isDeltaFile(f.getPath))
  .map(f => deltaVersion(f.getPath))
```

This PR adds overloads to those utility methods, which accept `FileStatus`. With that, the above code can simplify to:
```scala
listing
  .filter(isDeltaFile)
  .map(deltaVersion)
```

Existing unit tests cover all affected Delta code.

GitOrigin-RevId: 5602f22bccfe7585c43b0b20c265ff6904394aaa