Introduce a new file manifest for 'CONVERT TO DELTA' on c…
…atalog table

Currently, when _spark_metadata log is missing, "CONVERT TO DELTA" always scans the whole directory recursively for data files. However, this is not necessary for some cases. When the source table is a catalog table, we can use metadata from catalog to prune out-of-date files.

This PR introduces a new manifest to fetch data files based on partition informations from catalog, and it is invoked for the scenario above. The feature is protected with a SQL conf, whose default is true

Unit test is added in this PR.

GitOrigin-RevId: 0eb87398bc65e95d54ebf32c1bec2d1c683dd327