Remove duplicate rule ReplaceCurrentTimeInMergeInto
Remove the ReplaceCurrentTimeInMergeInto rule.
To have consistent timestamps in Delta, the same functionality had to be duplicated in PreprocessTableMerge. This is a follow up to remove the initial solution based on the ReplaceCurrentTimeInMergeInto rule. It also uses the code from the ComputeCurrentTime rule in PreprocessTableMerge when it is available.

Test coverage already there

GitOrigin-RevId: b848fe911479cbeb0a3949520dabf8434569382a