Fix ActionSerializerSuite
This test changes when/how actions are instantiated inside of ActionSerializerSuite when passed into testActionSerDe. We change the action type from Action to => Action. We also make the failing metadata lazy. This ensures that the metadata are instantiated after SparkFunSuite::beforeAll has run. This ensures that Utils.isTesting is true, meaning that the metadata will be instantiated with id testId.

Existing UTs.

Resolves delta-io/delta#1283
Closes delta-io/delta#1317

GitOrigin-RevId: 020565687710afa4aa240355ca4aca4c0ba9d4e8