Use ThreadUtils.parmap for optimize
## Description
Resolves https://github.com/delta-io/delta/issues/1220

Uses ThreadUtils.parmap to parallelize the compaction instead of parallel collections. This should improve the "tail" of the compaction execution. It seems to be that the parallel collections method buckets each job into maxThreads groups and then executes each group with one of the threads in the pool. I think this is more of a proper queue based approach so any remaining tasks can be done by any free thread.

Exsting UTs

## Does this PR introduce _any_ user-facing changes?
Just tail performance gain for optimize commands.

Closes delta-io/delta#1315

Signed-off-by: Scott Sandre <scott.sandre@databricks.com>
GitOrigin-RevId: 0ef683ac806153dd41c1cb74fc2fe1864dab1c76