Make test-only DeltaLog.forTable overloads implicit
Today, `DeltaLog.forTable` has a large number of overloads, most of which are only called from unit tests (i.e. passing custom clocks, or taking `java.io.File` as the table path). This PR moves all those methods to `DeltaTestImplicits`, so it is obvious which overloads are actually used in prod code.

If the code compiles, it should be correct. Meanwhile, the change almost exclusively affects unit tests.

GitOrigin-RevId: 5cc024c32f8680f248cb2b9807c6c538e9dcdea3