Change the status collection schema to follow the order in the table schema
Delta table collects column-level stats such as min/max and null counts. By default, stats for the first 32 columns are collected. However, the “first” is creating confusion for users. The public doc says Delta Lake collects statistics on the first delta.dataSkippingNumIndexedCols columns defined in your table schema, but the implementation collects the first columns based on the DataFrame schema for write. This PR proposes to change the stats collection to follow the order in the table schema.

Added new unit tests

GitOrigin-RevId: b852bcbd13462d72b68d4ef8223c5ee3e8344f9b