Refactor `snapshot` to `unsafeVolatileSnapshot`
Change the API `snapshot` to `unsafeVolatileSnapshot`. In all src code, we update the name to this new one to reflect the danger of using the api. In test code, we define an implicit helper method `snapshot` so that the test usages (of which there are hundreds) don't all have to change.

The reason for the change is that the `snapshot` def is backed by a volatile var. It has the potential to introduce race condition bugs since its value might change between sequential accesses.

There are a few cases where a better fix is available (for example, fetching the snapshot by doing a `deltaLog.update`), but in order to reduce the potential of bugs in such a large PR, I've left any actual logic changes to be done in follow-up PRs.

GitOrigin-RevId: 2313c46fb4c1db7d85e4632a75a34949f56aec2f