Add 'equivalent' for type comparison (#426)
Currently, comparison expressions fail with `IllegalArgumentException` for decimal type arguments that don’t have equal precision/scale. This should be valid (and is in Spark SQL). This impacts IN and all the binary comparison operators.

This change is to add `equivalent` method, to only be used internally by our code to compare data types.

Refer https://github.com/delta-io/connectors/issues/222 for detail.

Signed-off-by: Kwangin Jung <inylove82@gmail.com>