Use flag instead of extra method to keep the number of records when filtering files
Currently, there's ```filterFilesWithNumRecords``` and ```filterFiles``` to either filter with row level metrics or not. This is difficult to use as one needs to do an if statement to use the correct method. Instead we just unify those two methods and add a flag to indicate if we need row level metrics.

GitOrigin-RevId: 95f2497a55be5fba3782808b3143840ffb128911